Edited mainTexture Settings
Changed alphaIsTransparency to use the alpha channel of the .png file.
Sprite Mode to Multiple as this texture atlas as various type of
"textures" on it.
Pixels to Units to 200 ... to keep texture "fit" properly within
environment.
Wrap Mode to Clamp to prevent distortion or tiling artifacts from around
edges of sprites.